% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_acid_sequence.R
\name{a_acid_sequence}
\alias{a_acid_sequence}
\title{DNA to Amino acid sequence}
\usage{
a_acid_sequence(codons)
}
\arguments{
\item{codons}{string of a triplet of nucleotides in DNA that codes for an amino acid}
}
\value{
a peptide chain
}
\description{
Converting codons to amino acid sequence
}
