% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amino_acids}
\alias{amino_acids}
\title{World Health Organization TB data}
\format{
## `amino_acids`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{codons}{3 nucleotides related to an amino acid}
  ...
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
amino_acids
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
