% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_codons.R
\name{extract_codons}
\alias{extract_codons}
\title{Extract codons}
\usage{
extract_codons(input_sequence, start = 1)
}
\arguments{
\item{input_sequence}{input sequnce}

\item{start}{start}
}
\value{
codons
}
\description{
Extract codons
}
