% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dna.R
\name{generate_dna}
\alias{generate_dna}
\title{DNA nucleotides}
\usage{
generate_dna(length)
}
\arguments{
\item{length}{Length specifies how many nucleotides are added in a random order.}
}
\value{
Returns bases with no spaces and randomly picks nucleotides for creating DNA sequence.
}
\description{
DNA nucleotides
}
\examples{
generate_dna(5)

}
