% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_amino_acid_counts.R
\name{plot_amino_acid_counts}
\alias{plot_amino_acid_counts}
\title{Plot Amino Acid Counts}
\usage{
plot_amino_acid_counts(input_sequence)
}
\arguments{
\item{input_sequence}{The input sequence}
}
\value{
Amino Acid Plot
}
\description{
Plot Amino Acid Counts
}
