% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_content.R
\name{plot_content}
\alias{plot_content}
\title{Plot the content of each amino-acid in a protein}
\usage{
plot_content(protein)
}
\arguments{
\item{protein}{a sequence of letters(amino-acids)}
}
\value{
a plot with the amount of each amino-acid
}
\description{
Plot the content of each amino-acid in a protein
}
\examples{
plot_content("SPLIIFMRRE")
}
