% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rand_DNA_seq.R
\name{R_DNA_sequence}
\alias{R_DNA_sequence}
\title{Random DNA Sequence generator}
\usage{
R_DNA_sequence(Sequence_length)
}
\arguments{
\item{Sequence_length}{#Number of nucleotides the new DNA sequence should be}
}
\value{
Result is a random DNA sequence with N = 'Sequence_length' nucleotides
}
\description{
Random DNA Sequence generator
}
\examples{

X <- 10             #the length of the DNA sequence you want
R_DNA_sequence(X)   #Using the function on R



}
