% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_DNA.R
\name{random_DNA}
\alias{random_DNA}
\title{Sampling random DNA string}
\usage{
random_DNA(len = 10)
}
\arguments{
\item{len}{length of DNA sequence}
}
\value{
random DNA string
}
\description{
Sampling random DNA string
}
