% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_to_RNA.R
\name{RNA_seq}
\alias{RNA_seq}
\title{Converging DNA to RNA}
\usage{
RNA_seq(DNA_seq)
}
\arguments{
\item{DNA_seq}{the input DNA sequence}
}
\value{
the function gives the corresponding RNA sequence to the input DNA sequence
}
\description{
Converging DNA to RNA
}
