% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_split.R
\name{sequence_split}
\alias{sequence_split}
\title{Sequence splitter - Codon creator}
\usage{
sequence_split(sequence_base, start = 1)
}
\arguments{
\item{sequence_base}{The sequence of nucleotide bases given to the function}

\item{start}{The starting point for counting the codons}
}
\value{
Returns codons given from the base sequence
}
\description{
This function makes it so that you can split a base sequence into its codons.
}
