% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_into_codons.R
\name{split_into_codons}
\alias{split_into_codons}
\title{Split a sequence into codons}
\usage{
split_into_codons(sequence, start = 1)
}
\arguments{
\item{sequence}{The sequence to be split into codons. Should be a character vector}

\item{start}{The starting position. Positions before are not split into codons and not returned. Should be integer}
}
\value{
A character vector with codons
}
\description{
Split a sequence into codons
}
\examples{
split_into_codons("AUGUUUUAA")

}
