% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_replace.R
\name{transribe_dna_to_rna}
\alias{transribe_dna_to_rna}
\title{Match and replace}
\usage{
transcribe_dna_to_rna(dna_sequence)
}
\arguments{
\item{dna_sequence}{A sequence that is being matched.}
}
\value{
The function of the output is the replaced sequence that is made from param.
}
\description{
Match and replace
}
\examples{
transcribe_dna("ATATAT")
}
