% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRNA_rep.R
\name{tRNA_rep}
\alias{tRNA_rep}
\title{tRNA Replication}
\usage{
tRNA_rep(codons)
}
\arguments{
\item{codons}{3 letters (RNA nucleotides) sequence}
}
\value{
A letter which represents the amino acid the codon translates
}
\description{
tRNA Replication
}
\examples{
tRNA_rep("UUU")

}
