% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_amino_acid_counts.R
\name{plot_amino_acid_counts}
\alias{plot_amino_acid_counts}
\title{Plot Amino Acid Counts in a Sequence}
\usage{
plot_amino_acid_counts(amino_acid_sequence)
}
\arguments{
\item{amino_acid_sequence}{A character string representing an amino acid sequence.}
}
\value{
A ggplot object showing the frequency of each amino acid.
}
\description{
This function calculates the frequency of each amino acid in a given sequence and plots the counts as a bar chart.
}
