% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe_dna_to_rna.R
\name{transcribe_dna_to_rna}
\alias{transcribe_dna_to_rna}
\title{Transcribe DNA to RNA}
\usage{
transcribe_dna_to_rna(dna_sequence)
}
\arguments{
\item{dna_sequence}{A character string representing a DNA sequence.}
}
\value{
A character string representing the transcribed RNA sequence.
}
\description{
This function converts a DNA sequence to an RNA sequence by replacing all "T" nucleotides with "U".
}
